requirements.txt
# Core dependencies
numpy>=1.26.0
pillow>=10.0.0
opencv-python>=4.8.0

# Deep Learning frameworks
torch>=2.0.0
transformers>=4.45.0
accelerate>=0.20.0
einops>=0.6.0

# VLM API clients
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.25.0

# Configuration management
python-dotenv>=1.0.0

# Scientific computing
scipy>=1.11.0
matplotlib>=3.7.0
networkx>=3.0

# Data formats and I/O
PyYAML>=6.0
lxml>=4.9.0
xmlschema>=3.0.0

# HuggingFace ecosystem
huggingface-hub>=0.20.0
tokenizers>=0.15.0
safetensors>=0.4.0

# Additional utilities
requests>=2.31.0
tqdm>=4.65.0
pyclothoids>=0.2.0

# Qwen VL/Omni utilities
qwen-omni-utils

# Optional: Flash Attention (for faster inference with local VLMs)
# flash-attn>=2.0.0

# Optional: Qwen VL utilities (install if using Qwen VL models)
# qwen-vl-utils
